# [Perf] Map/MapWithDefault/OrderedSet allocation