--- a/packages/ember-metal/lib/map.js
+++ b/packages/ember-metal/lib/map.js
@@ -21,6 +21,7 @@
   `Ember.Map.create()` for symmetry with other Ember classes.
 */
 import { guidFor } from 'ember-utils';
+import { runInDebug } from 'ember-debug';
 
 function missingFunction(fn) {
   throw new TypeError(`${Object.prototype.toString.call(fn)} is not a function`);
@@ -41,17 +42,6 @@ function copyNull(obj) {
   return output;
 }
 
-function copyMap(original, newObject) {
-  let keys = original._keys.copy();
-  let values = copyNull(original._values);
-
-  newObject._keys = keys;
-  newObject._values = values;
-  newObject.size = original.size;
-
-  return newObject;
-}
-
 /**
   This class is used internally by Ember and Ember Data.
   Please do not use it at this time. We plan to clean it up
@@ -63,11 +53,12 @@ function copyMap(original, newObject) {
   @private
 */
 function OrderedSet() {
-  if (this instanceof OrderedSet) {
-    this.clear();
-  } else {
-    missingNew('OrderedSet');
-  }
+  runInDebug(() => {
+    if (!(this instanceof OrderedSet)) {
+      missingNew('OrderedSet');
+    }
+  })
+  this.clear();
 }
 
 /**
@@ -77,9 +68,7 @@ function OrderedSet() {
   @private
 */
 OrderedSet.create = function() {
-  let Constructor = this;
-
-  return new Constructor();
+  return new this();
 };
 
 OrderedSet.prototype = {
@@ -89,11 +78,18 @@ OrderedSet.prototype = {
     @private
   */
   clear() {
-    this.presenceSet = Object.create(null);
-    this.list = [];
+    this._presenceSet = undefined;
+    this._list = undefined;
     this.size = 0;
   },
 
+  get presenceSet() {
+    return this._presenceSet || (this._presenceSet = Object.create(null))
+  },
+
+  get list() {
+    return this._list || (this._list = []);
+  },
   /**
     @method add
     @param obj
@@ -123,6 +119,7 @@ OrderedSet.prototype = {
     @private
   */
   delete(obj, _guid) {
+    if (this.size === 0) { return;}
     let guid = _guid || guidFor(obj);
     let presenceSet = this.presenceSet;
     let list = this.list;
@@ -205,11 +202,11 @@ OrderedSet.prototype = {
     @private
   */
   copy() {
-    let Constructor = this.constructor;
-    let set = new Constructor();
+    let set = new this.constructor();
 
-    set.presenceSet = copyNull(this.presenceSet);
-    set.list = this.toArray();
+    if (this._presenceSet) { set._presenceSet = copyNull(this._presenceSet); }
+    if (this._list)        { set._list = this._list.slice(); }
+
     set.size = this.size;
 
     return set;
@@ -237,13 +234,12 @@ OrderedSet.prototype = {
   @constructor
 */
 function Map() {
-  if (this instanceof Map) {
-    this._keys = OrderedSet.create();
-    this._values = Object.create(null);
-    this.size = 0;
-  } else {
-    missingNew('Map');
-  }
+  runInDebug(() => {
+    if (!(this instanceof Map)) {
+      missingNew('Map');
+    }
+  });
+  this.clear();
 }
 
 /**
@@ -252,8 +248,7 @@ function Map() {
   @private
 */
 Map.create = function() {
-  let Constructor = this;
-  return new Constructor();
+  return new this();
 };
 
 Map.prototype = {
@@ -349,6 +344,7 @@ Map.prototype = {
     @private
   */
   has(key) {
+    if (this.size === 0) { return false; }
     return this._keys.has(key);
   },
 
@@ -390,18 +386,33 @@ Map.prototype = {
     @private
   */
   clear() {
-    this._keys.clear();
-    this._values = Object.create(null);
+    this.__keys = undefined;
+    this.__values = undefined;
     this.size = 0;
   },
 
+  get _keys() {
+    return this.__keys || (this.__keys = new OrderedSet());
+  },
+
+  get _values() {
+    return this.__values || (this.__values = Object.create(null));
+  },
   /**
     @method copy
     @return {Ember.Map}
     @private
   */
   copy() {
-    return copyMap(this, new Map());
+    let copied = new this.constructor();
+
+    if (this.size > 0) {
+      copied.__keys = this.__keys.copy();
+      copied.__values = copyNull(this.__values);
+      copied.size = this.size;
+    }
+
+    return copied;
   }
 };
 
@@ -416,7 +427,9 @@ Map.prototype = {
 */
 function MapWithDefault(options) {
   this._super$constructor();
-  this.defaultValue = options.defaultValue;
+  if (typeof options ==='object' && options !== null) {
+    this.defaultValue = options.defaultValue;
+  }
 }
 
 /**
@@ -440,6 +453,7 @@ MapWithDefault.prototype = Object.create(Map.prototype);
 MapWithDefault.prototype.constructor = MapWithDefault;
 MapWithDefault.prototype._super$constructor = Map;
 MapWithDefault.prototype._super$get = Map.prototype.get;
+MapWithDefault.prototype._super$copy = Map.prototype.copy;
 
 /**
   Retrieve the value associated with a given key.
@@ -467,10 +481,9 @@ MapWithDefault.prototype.get = function(key) {
   @private
 */
 MapWithDefault.prototype.copy = function() {
-  let Constructor = this.constructor;
-  return copyMap(this, new Constructor({
-    defaultValue: this.defaultValue
-  }));
+  let map = this._super$copy();
+  map.defaultValue = this.defaultValue;
+  return map;
 };
 
 export default Map;