@stefanpenner 2017-04-13
Pros:

* much faster OrderedSet and Map (in most scenarios)

Cons:

* more complexity/code