@runspired 2017-04-13
@stefanpenner thoughts on pulling this into a separate repo and deprecating the locations here?  If these things are not used by ember-core, and left here for ember-data, it would be better for ember-data to directly consume them. Especially as we iterate towards a goal of data working with stand-along glimmer.