@stefanpenner 2017-04-13
> @stefanpenner thoughts on pulling this into a separate repo and deprecating the locations here? If these things are not used by ember-core, and left here for ember-data, it would be better for ember-data to directly consume them.

I'm fine with making a second repo, I don't have the time to deprecate and all that fun stuff though.