@stefanpenner 2017-04-13
As it turns out i already have such a repo, this OrderedSet is faster then that one: https://github.com/stefanpenner/fast-ordered-set