@runspired 2017-04-13
@stefanpenner if you update that repo as well as landing this here, when I get a chance next week I'll PR the deprecations and the new location.