@runspired 2017-04-13
@stefanpenner I'll also deal with the deprecation RFC if there needs to be one, but given it's just a location change, do we need one?