@krisselden 2017-04-13
Do we not want to apply some of the same thing to the caches in Ember?  @rwjblue had a similar PR inspired by an adaptive map in lodash.