@stefanpenner 2017-04-13
> Do we not want to apply some of the same thing to the caches in Ember?

Thats possible, would we use OrderedSet directly or something more tailored to those use-cases? I suspect as Order doesn't matter, it may be something different anyways.

@krisselden Is the idea to replace the various lookup caches, path parsing caches etc?

---

It sounds like i'll make two repos, one for fast-ordered-set and one for fast-cache?