@stefanpenner 2017-04-14
We should test this in the wild some.

@runspired I believe you have a good scenario (what you have been doing for the other relationship tests?)