@stefanpenner 2017-04-24
Looks like ember-data monkey-patches OrderedSet and depends on `presenceSet` https://github.com/emberjs/data/blob/e5dac10e43f420bd3cd44e25f48b0295b93589fe/addon/-private/system/ordered-set.js#L21

- [ ] fix ember-data
- [ ] construct presentSet lazily on access for compatibility.