@locks 2018-09-10
Closing due to https://github.com/emberjs/rfcs/blob/b64b6df30483cb7a0ce39dc651d28b30a8b13bc3/text/0237-deprecation-ember-map.md.